+++ src/FunFair.BuildCheck.ProjectChecks/ReferencedPackages/ShouldUseAbstractionsLoggingPackage.cs
@@ -7,7 +7,7 @@ namespace FunFair.BuildCheck.ProjectChecks.ReferencedPackages
     ///     Checks that the logging abstractions package is used rather than the full version.
     /// </summary>
     [SuppressMessage(category: "ReSharper", checkId: "ClassNeverInstantiated.Global", Justification = "Created by DI")]
-    public sealed class ShouldUseAbstractionsLoggingPackage : ShouldUseAbstractionsPackage
+    public sealed class ShouldUseAbstractionsLoggingPackage : ShouldUseAlternatePackage
     {
         /// <summary>
         ///     Constructor.