+++ src/FunFair.BuildCheck.ProjectChecks/ReferencedPackages/ShouldUseAlternatePackage.cs
@@ -10,7 +10,7 @@ namespace FunFair.BuildCheck.ProjectChecks.ReferencedPackages
     /// <summary>
     ///     Checks to see if a non-'abstractions' package is used in a library rather than the slimmer, easier to update abstractions package.
     /// </summary>
-    public abstract class ShouldUseAbstractionsPackage : IProjectCheck
+    public abstract class ShouldUseAlternatePackage : IProjectCheck
     {
         private readonly ILogger _logger;
         private readonly string _matchPackageId;
@@ -22,7 +22,7 @@ namespace FunFair.BuildCheck.ProjectChecks.ReferencedPackages
         /// <param name="matchPackageId">The non-abstractions package.</param>
         /// <param name="usePackageId">The abstractions version of <paramref name="matchPackageId" />.</param>
         /// <param name="logger">Logging.</param>
-        protected ShouldUseAbstractionsPackage(string matchPackageId, string usePackageId, ILogger logger)
+        protected ShouldUseAlternatePackage(string matchPackageId, string usePackageId, ILogger logger)
         {
             this._matchPackageId = matchPackageId ?? throw new ArgumentNullException(nameof(matchPackageId));
             this._usePackageId = usePackageId ?? throw new ArgumentNullException(nameof(usePackageId));