+++ src/FunFair.BuildCheck.ProjectChecks/ReferencedPackages/ShouldUseFluentValidationAspNetCoreRatherThanFluentValidationPackage.cs
@@ -7,7 +7,7 @@ namespace FunFair.BuildCheck.ProjectChecks.ReferencedPackages
     ///     Checks that the correct version of FluentValidation is installed.
     /// </summary>
     [SuppressMessage(category: "ReSharper", checkId: "ClassNeverInstantiated.Global", Justification = "Created by DI")]
-    public sealed class ShouldUseFluentValidationAspNetCoreRatherThanFluentValidationPackage : ShouldUseAbstractionsPackage
+    public sealed class ShouldUseFluentValidationAspNetCoreRatherThanFluentValidationPackage : ShouldUseAlternatePackage
     {
         /// <summary>
         ///     Constructor.